# Optimization with "super particles"