--- a/pixi/src/main/java/org/openpixi/pixi/physics/Settings.java
+++ b/pixi/src/main/java/org/openpixi/pixi/physics/Settings.java
@@ -190,6 +190,10 @@ public class Settings {
 		return useGrid;
 	}
 
+	public int getNumOfThreads() {
+		return numOfThreads;
+	}
+
 	public ArrayList<IFieldGenerator> getFieldGenerators() {
 		return this.fieldGenerators;
 	}