--- a/pixi/src/main/java/org/openpixi/pixi/physics/SimulationType.java
+++ b/pixi/src/main/java/org/openpixi/pixi/physics/SimulationType.java
@@ -5,5 +5,6 @@ public enum SimulationType {
 	LorenzYangMills,
 	TemporalCGC,
 	TemporalCGCNGP,
+	TemporalOptimizedCGCNGP,
 	BoostInvariantCGC
 }