--- a/pixi/src/main/java/org/openpixi/pixi/physics/grid/CGCSuperParticleInterpolationNGP.java
+++ b/pixi/src/main/java/org/openpixi/pixi/physics/grid/CGCSuperParticleInterpolationNGP.java
@@ -1,5 +1,7 @@
 package org.openpixi.pixi.physics.grid;
 
+import org.openpixi.pixi.math.AlgebraElement;
+import org.openpixi.pixi.math.GroupElement;
 import org.openpixi.pixi.physics.particles.CGCSuperParticle;
 import org.openpixi.pixi.physics.particles.IParticle;
 
@@ -12,16 +14,31 @@ import org.openpixi.pixi.physics.particles.IParticle;
  */
 public class CGCSuperParticleInterpolationNGP implements  InterpolatorAlgorithm {
 	public void interpolateToGrid(IParticle p, Grid g) {
-		// Not implemented yet.
+		double at = g.getTemporalSpacing();
+		double as = g.getLatticeSpacing();
+
+		CGCSuperParticle P = (CGCSuperParticle) p;
+		if(P.needsUpdate(g.getSimulationSteps())) {
+			int indexOffset = P.getCurrentOffset(g.getSimulationSteps());
+			if(P.orientation > 0) {
+				for (int i = 0; i < P.numberOfParticles; i++) {
+					int index = i + indexOffset;
+					AlgebraElement J = P.Q[i].mult(as / at);
+					g.addJ(index, 0, J); // Optimizations only work for x-direction!
+					GroupElement U = g.getUnext(index, 0);
+					P.Q[i].actAssign(U.adj());
+				}
+			} else {
+				// Other orientation not yet implemented.
+			}
+		}
 	}
 
 	public void interpolateChargedensity(IParticle p, Grid g) {
-		// Not implemented yet.
 		CGCSuperParticle P = (CGCSuperParticle) p;
-		int indexOffset = P.indexOffset;
-		int ngpShift = (P.subLatticeShift < P.particlePerCell/2) ? 0 : P.particlesPerPlane;
+		int indexOffset = P.getCurrentNGPOffset(g.getSimulationSteps());
 		for (int i = 0; i < P.numberOfParticles; i++) {
-			int index = i + indexOffset + ngpShift;
+			int index = i + indexOffset;
 			g.addRho(index, P.Q[i]);
 		}
 	}