This pull request adds a new simulation type: _SimulationType.TemporalOptimizedCGCNGP_  or "temporal optimized cgc ngp" in YAML configurations.

Using this simulation type CGC simulations can be sped up and need less memory. Instead of having a large number of independent colored particles, particle charges are now grouped together to form a number of "super particles". Since the relative distances of the color charges in a single nucleus do not change during the simulation, we can exploit this fact to not only use less memory but also simplify the interpolation routines dramatically.

Notes:
- These very specific optimizations work only when the propagation direction of the initial conditions is set to 0.
- Due to a slightly altered method of cutting off negligible Gauss law violations in the initial conditions the numerical results of optimized simulations might not agree with the results of un-optimized simulations down to the last digit. 

### Benchmarks
Configuration: collision using incoherent MV model on a 128x64^2 grid.
System: i7 6700k (stock), 32gb

_Without optimizations_: 
Simulation time: 18 s (average **576ms**/step), heap dump* ~**1.32gb**
_Super particles_: 
Simulation time: 10 s (average **342ms**/step), heap dump* ~ **0.93gb**
_Super particles + SU(2) improvements_ #184 :
Simulation time: 6 s (average **204ms**/step), heap dump* ~ **0.69gb**

*) heap dump saved after initialization.